% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_colours.R
\name{scale_colour_dfe}
\alias{scale_colour_dfe}
\title{Add colour scale to ggplot with DfE theme}
\usage{
scale_colour_dfe(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The colour palette to use}

\item{discrete}{Is the scale discrete (defaults \code{TRUE})}

\item{reverse}{Reverse the order of the scale (defaults \code{FALSE})}

\item{...}{arguments passed on to \code{discrete_scale()}}
}
\value{
These are the palettes in their current form
\itemize{
\item \code{main} - Blue red turquoise (discrete, 3 colours)
\item \code{warm} - Red to purple (continuous)
\item \code{cool} - Blue to lime (continuous)
\item \code{full} - all colours in the brand (discrete, 7 colours)
\item \code{likert} - red to blue, grey midpoint (diverging)
\item \code{likert2} - green to purple, white midpoint (diverging)
\item \code{likert3} -  red to blue, white midpoint (diverging)
\item \code{heat} - white to red (continuous)
\item \code{heat2} - white to pink (continuous)
\item \code{cold} - white to blue (continuous)
\item \code{cold2} - white to turqouise (continuous)
}
}
\description{
Add colour scale to ggplot with DfE theme
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, colour = Species)) +
  geom_point() +
  scale_colour_dfe()

ggplot(iris, aes(x = Petal.Length, y = Sepal.Length,colour = Sepal.Width)) +
  geom_point(size = 3, alpha = 3/4) +
  scale_colour_dfe(palette = "heat", discrete = FALSE)


}
