% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_colours.R
\name{scale_fill_dfe}
\alias{scale_fill_dfe}
\title{Add fill scale to ggplot with DfE theme}
\usage{
scale_fill_dfe(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The colour palette to use}

\item{discrete}{Is the scale discrete (defaults TRUE)}

\item{reverse}{Reverse the order of the scale (defaults FALSE)}

\item{...}{arguments passed onto discrete_scale or scale_fill_gradientn}
}
\value{
These are the palettes in their current form \cr
main - Blue red turquoise (discrete, 3 colours) \cr
warm - Red to purple (continuous) \cr
cool - Blue to lime (continuous) \cr
full - all colours in the brand (discrete, 7 colours) \cr
likert - red to blue, grey midpoint (diverging) \cr
likert2 - green to purple, white midpoint (diverging) \cr
likert3 -  red to blue, white midpoint (diverging) \cr
heat - white to red (continuous) \cr
heat2 - white to pink (continuous) \cr
cold - white to blue (continuous) \cr
cold2 - white to turqouise (continuous)\cr
}
\description{
Add fill scale to ggplot with DfE theme
}
\examples{
iris \%>\%
 ggplot(aes(x=Species,y=Sepal.Length,fill=Species))+
 geom_col()+
 scale_fill_dfe()


data <- data.frame(murder = USArrests$Murder,
 state = tolower(rownames(USArrests)))
 map <- map_data("state")
 ggplot(data, aes(fill=murder))+
   geom_map(aes(map_id = state), map=map)+
   expand_limits(x = map$long, y=map$lat)+
   scale_fill_dfe(palette = 'heat',discrete = F)



}
