% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_colours.R
\name{scale_colour_dfe}
\alias{scale_colour_dfe}
\title{Add colour scale to ggplot with DfE theme}
\usage{
scale_colour_dfe(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The colour palette to use}

\item{discrete}{Is the scale discrete (defaults TRUE)}

\item{reverse}{Reverse the order of the scale (defaults FALSE)}

\item{...}{arguments passed onto discrete_scale}
}
\value{
These are the palettes in their current form \cr
main - Blue red turquoise (discrete, 3 colours) \cr
warm - Red to purple (continuous) \cr
cool - Blue to lime (continuous) \cr
full - all colours in the brand (discrete, 7 colours) \cr
likert - red to blue, grey midpoint (diverging) \cr
likert2 - green to purple, white midpoint (diverging) \cr
likert3 -  red to blue, white midpoint (diverging) \cr
heat - white to red (continuous) \cr
heat2 - white to pink (continuous) \cr
cold - white to blue (continuous) \cr
cold2 - white to turqouise (continuous)\cr
}
\description{
Add colour scale to ggplot with DfE theme
}
\examples{
library(ggplot2)
iris |>
ggplot(aes(x=Sepal.Width, y=Sepal.Length, colour = Species))+
geom_point()+
scale_colour_dfe()

iris |>
 ggplot(aes(x=Petal.Length,y=Sepal.Length,colour = Sepal.Width))+
 geom_point(size = 3, alpha = 3/4)+
 scale_colour_dfe(palette = 'heat', discrete = FALSE)


}
