% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_colours.R
\name{scale_fill_dfe}
\alias{scale_fill_dfe}
\title{Add fill scale to ggplot with DfE theme}
\usage{
scale_fill_dfe(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{The colour palette to use}

\item{discrete}{Is the scale discrete (defaults \code{TRUE})}

\item{reverse}{Reverse the order of the scale (defaults \code{FALSE})}

\item{...}{arguments passed onto discrete_scale or \code{scale_fill_gradientn()}}
}
\value{
These are the palettes in their current form
\itemize{
\item \code{main} - Blue red turquoise (discrete, 3 colours)
\item \code{warm} - Red to purple (continuous)
\item \code{cool} - Blue to lime (continuous)
\item \code{full} - all colours in the brand (discrete, 7 colours)
\item \code{likert} - red to blue, grey midpoint (diverging)
\item \code{likert2} - green to purple, white midpoint (diverging)
\item \code{likert3} -  red to blue, white midpoint (diverging)
\item \code{heat} - white to red (continuous)
\item \code{heat2} - white to pink (continuous)
\item \code{cold} - white to blue (continuous)
\item \code{cold2} - white to turqouise (continuous)
}
}
\description{
Add fill scale to ggplot with DfE theme
}
\examples{
library(ggplot2)

ggplot(iris, aes(Species, Sepal.Length, fill = Species)) +
  geom_col() +
  scale_fill_dfe()

ggplot(faithful, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_dfe(discrete = FALSE, palette = 'heat')

}
